requirements.txt
imageio==2.9.0
numpy>=1.18.0
pytorch_msssim==0.2.1
tqdm==4.36.1
torchvision>=0.9.0
scikit-image==0.16.2
torch>=1.8.0
PyYAML==5.4.1
tqdm
hydra-core==1.2.0
omegaconf==2.2.2
wandb
glob2
opencv-python==3.4.18.65
albumentations==1.0.0
adamp==0.3.0
adabelief-pytorch==0.2.1
segmentation-models-pytorch==0.2.1
#git+https://github.com/katsura-jp/pytorch-cosine-annealing-with-warmup
